// src/constants/colors.d.ts
export declare const COLORS: {
    secondary: {
        20: string;
        30: string;
        40: string;
        50: string;
        60: string;
        70: string;
        80: string;
        90: string;
        100: string;
    };
    primary: {
        5: string;
        10: string;
        20: string;
        30: string;
        40: string;
        50: string;
        60: string;
        70: string;
        80: string;
        90: string;
        100: string;
    };
    white: {
        10: string;
        20: string;
        50: string;
        100: string;
    };
    black: {
        100: string;
    };
    modalOpacity: {
        10: string;
        20: string;
    };
};